+++ Anna_Gamzinova/framework_from_scratch/CONSTANTS.py
