+++ Anna_Gamzinova/framework_from_scratch/conftest.py
@@ -1,9 +1,12 @@
 import pytest
+import json
 
-from Anna_Gamzinova.HW_18_framework_from_scratch.page_objects.main_page import MainPage
-from Anna_Gamzinova.HW_18_framework_from_scratch.page_objects.login_page import LoginPage
-from Anna_Gamzinova.HW_18_framework_from_scratch.utilities.config_parser import ReadConfig
-from Anna_Gamzinova.HW_18_framework_from_scratch.utilities.driver_factory import DriverFactory
+from Anna_Gamzinova.framework_from_scratch.CONSTANTS import ROOT_DIR
+from Anna_Gamzinova.framework_from_scratch.page_objects.main_page import MainPage
+from Anna_Gamzinova.framework_from_scratch.page_objects.login_page import LoginPage
+from Anna_Gamzinova.framework_from_scratch.utilities.config_parser import ReadConfig
+from Anna_Gamzinova.framework_from_scratch.utilities.configurations import Configurations
+from Anna_Gamzinova.framework_from_scratch.utilities.driver_factory import DriverFactory
 
 
 @pytest.fixture()
@@ -69,3 +72,10 @@ def open_overview_page(open_checkout_page):
     checkout_page.set_checkout_info(ReadConfig.get_first_name(), ReadConfig.get_last_name(), ReadConfig.get_zip())
     overview_page = checkout_page.click_continue_button()
     return overview_page
+
+
+@pytest.fixture()
+def env():
+    with open(f'{ROOT_DIR}/configurations/configurations.json') as file:
+        env_dict = json.loads(file.read())
+    return Configurations(**env_dict)
