+++ Anna_Gamzinova/framework_from_scratch/page_objects/finish_page.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
-from Anna_Gamzinova.HW_18_framework_from_scratch.utilities.web_ui.base_page import BasePage
-from Anna_Gamzinova.HW_18_framework_from_scratch.locators import xpath_locators
+from Anna_Gamzinova.framework_from_scratch.utilities.web_ui.base_page import BasePage
+from Anna_Gamzinova.framework_from_scratch.locators import xpath_locators
 
 
 class FinishPage(BasePage):
