+++ Anna_Gamzinova/framework_from_scratch/requirements.txt
@@ -1,3 +1,4 @@
 pytest==7.2.0
 selenium==4.6.0
 webdriver-manager==3.8.5
+json
