+++ Anna_Gamzinova/framework_from_scratch/utilities/web_ui/base_page.py
